+++ src/Factory.php
@@ -17,28 +17,12 @@ use Phlib\Beanstalk\Pool\SelectionStrategyInterface;
  */
 class Factory
 {
-    /**
-     * @deprecated 2.1.0 Static method will be changed to instance-based (with the same name) in the next major version
-     */
-    public static function create(string $host, int $port = Socket::DEFAULT_PORT, array $options = []): Connection
-    {
-        return (new static())->createBC($host, $port, $options);
-    }
-
-    public function createBC(string $host, int $port = Socket::DEFAULT_PORT, array $options = []): Connection
+    public function create(string $host, int $port = Socket::DEFAULT_PORT, array $options = []): Connection
     {
         return new Connection(new Socket($host, $port, $options));
     }
 
-    /**
-     * @deprecated 2.1.0 Static method will be changed to instance-based (with the same name) in the next major version
-     */
-    public static function createFromArray(array $config): ConnectionInterface
-    {
-        return (new static())->createFromArrayBC($config);
-    }
-
-    public function createFromArrayBC(array $config): ConnectionInterface
+    public function createFromArray(array $config): ConnectionInterface
     {
         if (array_key_exists('host', $config)) {
             $config = [
@@ -48,13 +32,13 @@ class Factory
 
         if (array_key_exists('server', $config)) {
             $server = $this->normalizeArgs($config['server']);
-            $connection = $this->createBC($server['host'], $server['port'], $server['options']);
+            $connection = $this->create($server['host'], $server['port'], $server['options']);
         } elseif (array_key_exists('servers', $config)) {
             if (!isset($config['strategyClass'])) {
                 $config['strategyClass'] = RoundRobinStrategy::class;
             }
-            $servers = $this->createConnectionsBC($config['servers']);
-            $strategy = $this->createStrategyBC($config['strategyClass']);
+            $servers = $this->createConnections($config['servers']);
+            $strategy = $this->createStrategy($config['strategyClass']);
             $connection = new Pool(new Collection($servers, $strategy));
         } else {
             throw new InvalidArgumentException('Missing required server(s) configuration');
@@ -64,18 +48,9 @@ class Factory
     }
 
     /**
-     * @deprecated 2.1.0 Static method will be changed to instance-based (with the same name) in the next major version
      * @return Connection[]
      */
-    public static function createConnections(array $servers): array
-    {
-        return (new static())->createConnectionsBC($servers);
-    }
-
-    /**
-     * @return Connection[]
-     */
-    public function createConnectionsBC(array $servers): array
+    public function createConnections(array $servers): array
     {
         $connections = [];
         foreach ($servers as $server) {
@@ -83,21 +58,13 @@ class Factory
                 continue;
             }
             $server = $this->normalizeArgs($server);
-            $connection = $this->createBC($server['host'], $server['port'], $server['options']);
+            $connection = $this->create($server['host'], $server['port'], $server['options']);
             $connections[] = $connection;
         }
         return $connections;
     }
 
-    /**
-     * @deprecated 2.1.0 Static method will be changed to instance-based (with the same name) in the next major version
-     */
-    public static function createStrategy(string $class): SelectionStrategyInterface
-    {
-        return (new static())->createStrategyBC($class);
-    }
-
-    public function createStrategyBC(string $class): SelectionStrategyInterface
+    public function createStrategy(string $class): SelectionStrategyInterface
     {
         if (!class_exists($class)) {
             throw new InvalidArgumentException("Specified Pool strategy class '{$class}' does not exist.");