+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Changed
+- **BC break**: Deprecated static Factory methods are now instance-based.
 
 ## [2.1.0] - 2023-07-05
 ### Added